+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -14,6 +14,7 @@ import fr.maxlego08.essentials.module.modules.customcommands.CustomCommandModule
 import dev.yanianz.essentials.customscreens.CustomScreenModule;
 import dev.yanianz.essentials.polls.PollsModule;
 import dev.yanianz.essentials.nicknames.NicknamesModule;
+import dev.yanianz.essentials.bubbles.ChatBubblesModule;
 import dev.yanianz.essentials.chatcustomization.ChatCustomizationModule;
 import dev.yanianz.essentials.reports.ReportsModule;
 import dev.yanianz.essentials.notes.NotesModule;
@@ -103,6 +104,7 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(NicknamesModule.class, new NicknamesModule(this.plugin));
         this.modules.put(ReportsModule.class, new ReportsModule(this.plugin));
         this.modules.put(ChatCustomizationModule.class, new ChatCustomizationModule(this.plugin));
+        this.modules.put(ChatBubblesModule.class, new ChatBubblesModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }